adjust alert to sweet alert, rewrite using flow, such as login, create cocktail, comment, upload head pic

+++ public/js/login.js
@@ -7,7 +7,6 @@ if(accessToken&&user_info){
 
 
 
-
 $('.form').find('input, textarea').on('keyup blur focus', function (e) {
 
     var $this = $(this),
@@ -57,7 +56,17 @@ function sendSignUpData() {
     let name = document.getElementById('signup_name').value;
     let email = document.getElementById('signup_email').value;
     let password = document.getElementById('signup_password').value;
-    fetch(url, {
+
+    if(!name||!email||!password){
+        Swal.fire({
+            title: 'Please input all correctly !',
+            text: 'Do you want to continue?',
+            icon: 'warning',
+            confirmButtonText: 'OK'
+        });
+        return;
+    }
+    {fetch(url, {
         method: 'POST',
         body: JSON.stringify({
             query : `mutation createser($email: String!, $password: String!,$name:String!) {
@@ -79,9 +88,29 @@ function sendSignUpData() {
     })
         .then(res => res.json())
         .then(result => {
+            if(result.errors){
+
+
+                Swal.fire({
+                    title:  result.errors[0].message,
+                    text: 'Please try again!',
+                    icon: 'error',
+                    confirmButtonText: 'OK'
+                });
+                return;
+            }
+
+            console.log(result);
             if(result.data.createUser){
 
-                alert(`Welcome come! ${name}`);
+                Swal.fire({
+                    title: `Welcome come! ${name}! `,
+                    text: 'Enjoy tonight!',
+                    icon: 'success',
+                    showConfirmButton: false,
+                    timer: 2000,
+                    timerProgressBar: true,
+                });
                 let {accessToken,id} = result.data.createUser;
                 let user = {accessToken,id , photo :'', gender:'' };
                 localStorage.setItem('accessToken', JSON.stringify(accessToken));
@@ -90,17 +119,35 @@ function sendSignUpData() {
                 // localStorage.setItem('user_info', JSON.stringify(result.data.user));
                 // location.reload();
             } else {
-                alert(result.error);
-                // location.reload();
+
+                Swal.fire({
+                    title: 'Error!',
+                    text: `${result.error}`,
+                    icon: 'error',
+                    confirmButtonText: 'Try again!'
+                });
             }
 
 
-        });
+        });}
 }
+
+
 
 function sendLogINData(){
     let email = document.getElementById('signin_email').value;
     let password = document.getElementById('signin_password').value;
+
+    if(!email||!password){
+        Swal.fire({
+            title: 'Error!',
+            text: 'Should not be empty',
+            icon: 'error',
+            confirmButtonText: 'Try again!'
+        });
+        return;
+    }
+
     fetch(url, {
         method: 'POST',
         body: JSON.stringify({
@@ -124,7 +171,7 @@ function sendLogINData(){
         }
     })
         .then(res => res.json())
-        .then(user_info => {
+        .then( async (user_info) =>  {
             if (user_info.data.login) {
                 let {accessToken, name, id } = user_info.data.login;
 
@@ -132,14 +179,29 @@ function sendLogINData(){
                 //     id,
                 //     name,
                 // };
-                alert(`Welcome come! ${name}`);
+                await Swal.fire({
+                    title: `Welcome ! ${name}`,
+                    text: 'Enjoy tonight!',
+                    icon: 'success',
+                    showConfirmButton: false,
+                    timer: 2000,
+                    timerProgressBar: true,
+                });
+
+
+
                 localStorage.setItem('accessToken', JSON.stringify(accessToken));
                 // localStorage.setItem('user_info', JSON.stringify(user));
                 window.location.replace('/profile.html');
 
             } else {
-                alert('Wrong email or password, Please try again.');
-                location.reload();
+                Swal.fire({
+                    title: 'Error!',
+                    text: 'Wrong E-mail or password',
+                    icon: 'error',
+                    confirmButtonText: 'Try again!'
+                });
+                // location.reload();
             }
 
 
@@ -157,3 +219,4 @@ function logout(){
     }
 
 }
+
